functions/send_encounters/requirements.txt
pytz==2019.3
google-cloud-bigquery==1.24.0
google-cloud-datastore==1.11.0